Delete unused models and update repository methods
Removed several obsolete models from the WalletService.Data project. Updated repository methods and service interfaces to replace type 'int' with 'long' for IDs and revised method signatures accordingly. Updated database connection strings to use PostgreSQL instead of SQL Server.

+++ WalletService.Core/PaymentStrategies/BalancePaymentStrategy.cs
@@ -46,7 +46,7 @@ public class BalancePaymentStrategy : IBalancePaymentStrategy
         _bonusRepository         = bonusRepository;
     }
 
-    private async Task<BalanceInformationDto> GetBalanceInformationByAffiliateId(int affiliateId,int brandId)
+    private async Task<BalanceInformationDto> GetBalanceInformationByAffiliateId(int affiliateId,long brandId)
     {
         var amountRequests    = await _walletRequestRepository.GetTotalWalletRequestAmountByAffiliateId(affiliateId, brandId);
         var availableBalance  = await _walletRepository.GetAvailableBalanceByAffiliateId(affiliateId,brandId);
