+++ WalletService.Data/Adapters/AccountServiceAdapter.cs
@@ -21,7 +21,7 @@ public class AccountServiceAdapter : BaseAdapter, IAccountServiceAdapter
     protected override string? GetTokenUrl()
         => AppSettings.EndpointTokens?.AccountServiceToken;
     
-    protected override string? GetWebToken(int brandId)
+    protected override string? GetWebToken(long brandId)
     {
         return brandId switch
         {
@@ -31,7 +31,7 @@ public class AccountServiceAdapter : BaseAdapter, IAccountServiceAdapter
         };
     }
 
-    public Task<IRestResponse> VerificationCode(string code, string password, int userId, int brandId)
+    public Task<IRestResponse> VerificationCode(string code, string password, int userId, long brandId)
     {
         var requestValidationCode = new RequestValidationCode
         {
@@ -43,7 +43,7 @@ public class AccountServiceAdapter : BaseAdapter, IAccountServiceAdapter
         return Post("/userAffiliateInfo/validationCode/", requestValidationCode.ToJsonString(),brandId);
     }
 
-    public Task<IRestResponse> GetAccountsToEcoPool(int[] id, int levels, int brandId)
+    public Task<IRestResponse> GetAccountsToEcoPool(int[] id, int levels, long brandId)
     {
         var json = new
         {
@@ -54,21 +54,21 @@ public class AccountServiceAdapter : BaseAdapter, IAccountServiceAdapter
         return Post($"/userAffiliateInfo/get_accounts_eco_pool/", json, brandId);
     }
 
-    public Task<IRestResponse> UpdateActivationDate(int id, int brandId)
+    public Task<IRestResponse> UpdateActivationDate(int id, long brandId)
     {
         return Put($"/userAffiliateInfo/update_activation_date/{id}/", brandId);
     }
 
-    public Task<IRestResponse> RevertActivationUser(int id, int brandId)
+    public Task<IRestResponse> RevertActivationUser(int id, long brandId)
     {
         return Put($"/userAffiliateInfo/revert_activation/{id}/", brandId);
     }
-    public Task<IRestResponse> GetTotalActiveMembers(int brandId)
+    public Task<IRestResponse> GetTotalActiveMembers(long brandId)
     {
         return Get($"/userAffiliateInfo/getTotalActiveMembers/", new Dictionary<string, string>(), brandId);
     }
     
-    public async Task<IRestResponse> UpdateGradingByUser(int userId, int gradingId, int brandId)
+    public async Task<IRestResponse> UpdateGradingByUser(int userId, int gradingId, long brandId)
     {
         try
         {
@@ -82,7 +82,7 @@ public class AccountServiceAdapter : BaseAdapter, IAccountServiceAdapter
         }
     }
 
-    public Task<IRestResponse> GetHave2Children(int[]                       users, int brandId)
+    public Task<IRestResponse> GetHave2Children(int[]                       users, long brandId)
     {
         var data = new Dictionary<string, object>
         {
@@ -92,33 +92,33 @@ public class AccountServiceAdapter : BaseAdapter, IAccountServiceAdapter
         return Post($"/matrix/have_2_children/", data, brandId);
     }
     
-    public Task<IRestResponse> GetTreeModel4(Dictionary<int, decimal>       dictionary, int brandId)
+    public Task<IRestResponse> GetTreeModel4(Dictionary<int, decimal>       dictionary, long brandId)
     {
         var json = dictionary.ToJsonString();
         return Post($"/leaderboard/model4/getResultTree/", json, brandId);
     }
     
-    public Task<IRestResponse> AddTreeModel5(IEnumerable<LeaderBoardModel5> leaderBoard, int brandId)
+    public Task<IRestResponse> AddTreeModel5(IEnumerable<LeaderBoardModel5> leaderBoard, long brandId)
     {
         return Post($"/leaderboard/model5/addTree/", leaderBoard.ToJsonString(), brandId);
     }
     
-    public Task<IRestResponse> AddTreeModel6(IEnumerable<LeaderBoardModel6> leaderBoard , int brandId)
+    public Task<IRestResponse> AddTreeModel6(IEnumerable<LeaderBoardModel6> leaderBoard , long brandId)
     {
         return Post($"/leaderboard/model6/addTree/", leaderBoard.ToJsonString(), brandId);
     }
 
-    public Task<IRestResponse> DeleteTreeModel6(int brandId)
+    public Task<IRestResponse> DeleteTreeModel6(long brandId)
     {
         return Post($"/leaderboard/model6/deleteTree/", new Dictionary<string,string>().ToJsonString(), brandId);
     }
 
-    public Task<IRestResponse> DeleteTreeModel5(int brandId)
+    public Task<IRestResponse> DeleteTreeModel5(long brandId)
     {
         return Post($"/leaderboard/model5/deleteTree/", new Dictionary<string,string>().ToJsonString(), brandId);
     }
 
-    public async Task<UserInfoResponse?> GetUserInfo(int id, int brandId)
+    public async Task<UserInfoResponse?> GetUserInfo(int id, long brandId)
     {
         var response = await Get($"/userAffiliateInfo/get_user_id/{id}/", new Dictionary<string, string>(), brandId);
         if (!response.IsSuccessful)
@@ -150,7 +150,7 @@ public class AccountServiceAdapter : BaseAdapter, IAccountServiceAdapter
         return userInfoResponse;
     }
     
-    public async Task<NetworkDetailsResponse> NetworkDetails(int id, int brandId)
+    public async Task<NetworkDetailsResponse> NetworkDetails(int id, long brandId)
     {
         var response = await Get($"/userAffiliateInfo/getNetworkDetails/{id}", new Dictionary<string, string>(), brandId);
         if (!response.IsSuccessful)
@@ -164,12 +164,12 @@ public class AccountServiceAdapter : BaseAdapter, IAccountServiceAdapter
         return networkDetails;
     }
     
-    public Task<IRestResponse> GetAffiliateByUserName(string userName, int brandId)
+    public Task<IRestResponse> GetAffiliateByUserName(string userName, long brandId)
         => Get($"/userAffiliateInfo/get_user_username/{userName}/", new Dictionary<string, string>(), brandId);
 
-    public Task<IRestResponse> GetPersonalNetwork(int id, int brandId)
+    public Task<IRestResponse> GetPersonalNetwork(int id, long brandId)
         => Get($"/userAffiliateInfo/getPersonalNetwork/{id}/", new Dictionary<string, string>(), brandId);
     
-    public Task<IRestResponse> GetAffiliateBtcByAffiliateId(int affiliateId, int brandId)
+    public Task<IRestResponse> GetAffiliateBtcByAffiliateId(int affiliateId, long brandId)
         => Get($"/AffiliateBtc/get_affiliate_btc_by_affiliate_id/{affiliateId.ToJsonString()}/", new Dictionary<string, string>(), brandId);
 }
