+++ WalletService.Models/Requests/EcoPoolConfigurationRequest/EcoPoolConfigurationRequest.cs
@@ -3,11 +3,11 @@ namespace WalletService.Models.Requests.EcoPoolConfigurationRequest;
 public class EcoPoolConfigurationRequest
 {
     public int? Id { get; set; }
-    public double CompanyPercentage { get; set; }
+    public decimal CompanyPercentage { get; set; }
     
-    public double CompanyPercentageLevels { get; set; }
-    public double EcoPoolPercentage { get; set; }
-    public double MaxGainLimit { get; set; }
+    public decimal CompanyPercentageLevels { get; set; }
+    public decimal EcoPoolPercentage { get; set; }
+    public decimal MaxGainLimit { get; set; }
     public DateTime DateInit { get; set; }
     public DateTime DateEnd { get; set; }
     public int Case { get; set; }
