+++ WalletService.Models/Requests/WalletRequest/InvoiceDetailsTransactionRequest.cs
@@ -23,7 +23,7 @@ public class InvoiceDetailsTransactionRequest
     
     public int? DaysToPayQuantity { get; set; }
     public byte ProductStart { get; set; }
-    public int BrandId { get; set; }
+    public long BrandId { get; set; }
     
     
 }
